+++ src/constants/panel.ts
@@ -13,3 +13,19 @@ export const StickyOptions = [
   { value: true, label: 'Enabled', description: 'Follow when scrolling.' },
   { value: false, label: 'Disabled', description: 'Disable sticky position.' },
 ];
+
+/**
+ * Filter Options
+ */
+export const FilterOptions = [
+  { value: true, label: 'Enabled', description: 'Display Table Filter.' },
+  { value: false, label: 'Disabled', description: 'Hide Table Filter.' },
+];
+
+/**
+ * Favorites Options
+ */
+export const FavoritesOptions = [
+  { value: true, label: 'Enabled', description: 'Allows selecting and filtering favorites.' },
+  { value: false, label: 'Disabled', description: 'Hide favorites.' },
+];