+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Change Log
 
-## 1.3.0 (IN PROGRESS)
+## 1.3.0 (2023-06-30)
+
+### Features / Enhancements
+
+- Add Tutorial (#21)
+- Update adding new Level to last (#22)
+
+## 1.3.0 (2023-06-30)
 
 ### Features / Enhancements
 
 - Add Display Header option (#15)
 - Add the ability to collapse parent rows (#16)
 - Add a Sticky option to follow when scrolling (#17)
+- Add Table Filter (#18)
+- Add selecting favorites (#19)
+- Update documentation (#20)
 
 ## 1.2.0 (2023-06-26)